Added another config function, fixed no config = no dbs bug, and others
Added config function 'prepConfig()' which initializes the database and table.
Fixed a bug where when no config existed or the flag 'show_system_tables' was missing caused none of the databases to load.
Embedded settings into the view of the configuration whenever the table is loaded or settings is clicked in the User dropdown.